% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_config.R
\name{cfg_params}
\alias{cfg_params}
\title{Get the \code{params} section with defaults}
\usage{
cfg_params(cfg, defaults = list())
}
\arguments{
\item{cfg}{config list}

\item{defaults}{named list of default values}
}
\value{
named list
}
\description{
Get the \code{params} section with defaults
}
