% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_config.R
\name{cfg_paths}
\alias{cfg_paths}
\title{Get the \code{paths} section with optional required-key checks}
\usage{
cfg_paths(cfg, required = NULL)
}
\arguments{
\item{cfg}{config list}

\item{required}{character vector of keys that must exist}
}
\value{
named list
}
\description{
Get the \code{paths} section with optional required-key checks
}
