% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_config.R
\name{io_resolve_path}
\alias{io_resolve_path}
\title{Normalize/resolve a path (tilde-expand, absolute if possible)}
\usage{
io_resolve_path(x)
}
\description{
Normalize/resolve a path (tilde-expand, absolute if possible)
}
\keyword{internal}
