% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_config.R
\name{io_save_table}
\alias{io_save_table}
\title{Save a data.frame safely to CSV in an output directory}
\usage{
io_save_table(df, out_dir, filename)
}
\description{
Save a data.frame safely to CSV in an output directory
}
\keyword{internal}
