% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{remove_pos}
\alias{remove_pos}
\title{Remove drugs that show any positive CMap score}
\usage{
remove_pos(x)
}
\arguments{
\item{x}{Matrix/data frame of CMap scores (rows = drugs, cols = datasets).}
}
\value{
The filtered matrix/data frame with only non-positive rows kept.
}
\description{
Remove drugs that show any positive CMap score
}
