% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{dr_cli}
\alias{dr_cli}
\title{Command-Line Interface}
\usage{
dr_cli()
}
\description{
Provides the DRPipe CLI entry point for running drug repurposing pipeline and analysis
from the command line. Supports configuration via YAML files and enables
batch processing with various output options.
}
