% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chembl_validation.R
\name{load_chembl_chemreps}
\alias{load_chembl_chemreps}
\title{Load ChEMBL Chemical Representations}
\usage{
load_chembl_chemreps(chemreps_path, verbose = TRUE)
}
\arguments{
\item{chemreps_path}{Path to chembl_36_chemreps.txt.gz (gzipped tab-separated)}

\item{verbose}{Print progress messages.}
}
\value{
data.frame with drug structures mapped to chembl_id/molregno
}
\description{
Load ChEMBL Chemical Representations
}
\examples{
# chemreps <- load_chembl_chemreps("path/to/chembl_36_chemreps.txt.gz")

}
